Move InStock Button to its own component
